Implements cart dialog
Replaces the placeholder text in the cart dialog with a dynamic form to display and edit cart details.

Introduces a standalone CartFormComponent to encapsulate the cart form logic, including form fields for title, note, total, online purchase status, delivery status, and dates.

Adds a seller selector component to choose a supplier for the cart.

Adds filter pipe to filter the seller list.
#TODO  complete seller selector and complete the powers