Enhances seller selection in cart form
Improves the seller selection process within the cart form by integrating a bottom sheet selector.

Adds filtering by name and distance.

This enhances usability and provides a more streamlined experience for users when choosing a seller for their cart.#TODO how to inform the form that we set the seller